unified-quant-platform/requirements_v2.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.3
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
pandas==2.0.0
numpy==1.24.3
PyYAML==6.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2

# 数据库驱动（正式环境使用 MySQL）
# PyMySQL==1.1.0  # 测试环境使用 SQLite
